(PC-4962) : Rework the login use case
