+++ src/pcapi/core/users/exceptions.py
@@ -1,0 +1,14 @@
+class CredentialsException(Exception):
+    pass
+
+
+class InvalidIdentifier(CredentialsException):
+    pass
+
+
+class InvalidPassword(CredentialsException):
+    pass
+
+
+class UnvalidatedAccount(CredentialsException):
+    pass
