+++ src/pcapi/routes/beneficiaries.py
@@ -4,13 +4,15 @@ from flask_login import current_user, \
     login_required, \
     login_user
 
+from pcapi.core.users import api as user_api
+from pcapi.core.users import exceptions as user_exceptions
+from pcapi.domain.beneficiary.beneficiary_licence import is_licence_token_valid
 from pcapi.flask_app import private_api, \
     public_api
-from pcapi.domain.beneficiary.beneficiary_licence import is_licence_token_valid
+from pcapi.models.api_errors import ApiErrors
 from pcapi.routes.serialization import as_dict
 from pcapi.use_cases.update_user_informations import update_user_informations, \
     AlterableUserInformations
-from pcapi.utils.credentials import get_user_with_credentials
 from pcapi.utils.includes import BENEFICIARY_INCLUDES
 from pcapi.utils.login_manager import stamp_session
 from pcapi.utils.rest import expect_json_data, \
@@ -62,7 +64,19 @@ def signin_beneficiary():
     identifier = json.get("identifier")
     password = json.get("password")
     check_valid_signin(identifier, password)
-    user = get_user_with_credentials(identifier, password)
+    errors = ApiErrors()
+    errors.status_code = 401
+    try:
+        user = user_api.get_user_with_credentials(identifier, password)
+    except user_exceptions.InvalidIdentifier as exc:
+        errors.add_error('identifier', 'Identifiant incorrect')
+        raise errors from exc
+    except user_exceptions.UnvalidatedAccount as exc:
+        errors.add_error('identifier', "Ce compte n'est pas validé.")
+        raise errors from exc
+    except user_exceptions.InvalidPassword as exc:
+        errors.add_error('password', 'Mot de passe incorrect')
+        raise errors from exc
     login_user(user, remember=True)
     stamp_session(user)
     return jsonify(), 200
