+++ src/pcapi/routes/native/v1/authentication.py
@@ -3,9 +3,10 @@ from flask_jwt_extended import (
     get_jwt_identity
 )
 
+from pcapi.core.users import api as user_api
+from pcapi.core.users import exceptions as user_exceptions
 from pcapi.models.api_errors import ApiErrors
 from pcapi.serialization.decorator import spectree_serialize
-from pcapi.utils.credentials import get_user_with_credentials
 
 from . import blueprint
 from .serialization import authentication
@@ -15,12 +16,12 @@ from .serialization import authentication
 @spectree_serialize(response_model=authentication.SigninResponse, on_success_status=200, api=blueprint.api)  # type: ignore
 def signin(body: authentication.SigninRequest) -> authentication.SigninResponse:
     try:
-        get_user_with_credentials(body.identifier, body.password)
-    except ApiErrors as original_errors:
+        user_api.get_user_with_credentials(body.identifier, body.password)
+    except user_exceptions.CredentialsException as exc:
         errors = ApiErrors()
         errors.status_code = 400
         errors.add_error('general', 'Identifiant ou Mot de passe incorrect')
-        raise errors from original_errors
+        raise errors from exc
     # TODO: Should we fill the token with some less trivial data ?
     return authentication.SigninResponse(
         access_token=create_access_token(identity=body.identifier),
