+++ src/pcapi/utils/login_manager.py
@@ -4,10 +4,11 @@ import uuid
 from flask import current_app as app, jsonify, session
 from flask_login import login_user
 
+from pcapi.core.users import api as user_api
+from pcapi.core.users import exceptions as user_exceptions
 from pcapi.models.api_errors import ApiErrors
 from pcapi.models.user_sql_entity import UserSQLEntity
 from pcapi.repository.user_session_queries import existing_user_session, register_user_session, delete_user_session
-from pcapi.utils.credentials import get_user_with_credentials
 
 
 @app.login_manager.user_loader
@@ -25,7 +26,19 @@ def get_user_with_request(request):
     auth = request.authorization
     if not auth:
         return None
-    user = get_user_with_credentials(auth.username, auth.password)
+    errors = ApiErrors()
+    errors.status_code = 401
+    try:
+        user = user_api.get_user_with_credentials(auth.username, auth.password)
+    except user_exceptions.InvalidIdentifier as exc:
+        errors.add_error('identifier', 'Identifiant incorrect')
+        raise errors from exc
+    except user_exceptions.UnvalidatedAccount as exc:
+        errors.add_error('identifier', "Ce compte n'est pas validé.")
+        raise errors from exc
+    except user_exceptions.InvalidPassword as exc:
+        errors.add_error('password', 'Mot de passe incorrect')
+        raise errors from exc
     login_user(user, remember=True)
     stamp_session(user)
     return user
